requirements.txt
streamlit
openai
langchain
langchain-openai
langchain-community
python-dotenv
pypdf

